requirements/reuqirements.example.txt
asgiref==3.2.7
bcrypt==3.1.7
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.9.2
Django==2.2.13
django-environ==0.4.5
django-jet==1.0.8
django-js-asset==1.2.2
django-modeltranslation==0.15
django-nested-inline==0.4.2
django-rosetta==0.9.4
djangorestframework==3.11.0
drf-yasg==1.17.1
Fabric3==1.13.1.post1
fcm-django==0.3.4
idna==2.9
inflection==0.5.0
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
marshmallow==3.6.1
packaging==20.4
paramiko==2.7.1
polib==1.1.0
psycopg2-binary==2.8.5
pycparser==2.20
pyfcm==1.4.7
PyNaCl==1.4.0
pyparsing==2.4.7
python-dotenv==0.13.0
pytz==2020.1
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.15.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.9
